feat (gameplay) : modify loop to create input instead of div, config input for write letter.
I modify the loop who generated div depend of length of name to generate input instead of div. With this modify, we can write on the line.
Add an eventListener on input, to center the text.

Some trouble with junction of name, we can't choose the input that we select.
For moment, fix this problem with an opacity on input.